local-requirements.txt
flask==2.3.3
flask-login==0.6.2
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
pymysql==1.1.0
email-validator==2.1.0
markdown==3.5
markupsafe==2.1.3
werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
